Adds VSCode-related settings (#30)
VSCode does not currently support nested Git repositories that are not
submodules. A work-around is the use of a .code-workspace file, which
has been added in this commit. See
https://github.com/microsoft/vscode/issues/37947 for details. The
.gitignore file has also had "**/.vscode/" added to allow VSCode users
to configure additional personal settings.

**Note:** This is related to pull requests in all other kazoo5 repos,
but there are no specific dependencies between them. This PR is the
one that is most helpful for VSCode users.