Add retry mechanism for GIT commands (#68)
Days like today I don't know if it is my ISP or GitHub, but I am
getting a lot of timeouts whenever I try to update kazoo-*
repositories, same was happening when pinging github.com, 1 out of 3
tries was working, here is the output I was seeing a lot of times for
most of the repositories:

```bash
ssh: connect to host github.com port 22: Connection timed out
fatal: Could not read from remote repository.

Please make sure you have the correct access rights and the repository exists.
```

So this PR aims to add a retry mechanism for GIT commands for days
like today.