Ensure dependent apps are built already (#71)
When one KAZOO app depends on another (usually via an "integration"
like a crossbar module, webhook, etc) to exist, we need to make sure
the other app is built. This is particularly important if the module
uses a behaviour defined in the other app.

We also make sure the ebin/ of all dependent apps are listed in the
erlc command's -pa section.