Updated and enhanced elvis usage. (#73)
The elvis formatting checker was not running. The version was too
old. The dep has been updated to the most recent stable version.

This does mean that any system that wishes to use elvis will need 7z
installed on it. For centos, this means install p7zip-plugins and
p7zip.

To further enhance the use of elvis, a custom formating checker has
been added. It checks that all "raw" atoms are quoted. Raw in this
case means atoms that are not used as record field names, function
names, module names, or map keys. It lives in
`kazoo_stdlib/src/kz_style.erl`.