Use immediate expansion for variable ROOT in makefile (#90)
I'm not expert in Makefiles but I'm seeing any version of `make` on
macOS, be it BSD or GNU Make) have issue with expanding `ROOT`. I
think more specifically if we use the `ROOT` inside another make
`shell` function it is not expanded. Other usage of `ROOT` in regular
assignments are unaffected.

Like in `CHANGED_SWAGGER`, the `ROOT` won't be expanded so the shell
tries to run `/kgit` which does not exists, hence on any run of `make`
you see this line printed repeatedly:

```
make: /kgit: No such file or directory
```

By simply using `:=` in ROOT assignment, we make it expanded once and
avoid this issue.